Move custom exercise form into separate component
